Use consistent module name - separate words with underscore.
