+++ st2api/st2api/controllers/v1/rule_views.py
