+++ st2common/st2common/openapi.yaml
@@ -422,7 +422,7 @@ paths:
             $ref: '#/definitions/Error'
   /api/v1/actions/views/parameters/{action_id}:
     get:
-      operationId: st2api.controllers.v1.actionviews:parameters_view_controller.get_one
+      operationId: st2api.controllers.v1.action_views:parameters_view_controller.get_one
       description: |
         Get parameters for an action.
       parameters:
@@ -453,7 +453,7 @@ paths:
             $ref: '#/definitions/Error'
   /api/v1/actions/views/overview:
     get:
-      operationId: st2api.controllers.v1.actionviews:overview_controller.get_all
+      operationId: st2api.controllers.v1.action_views:overview_controller.get_all
       x-permissions: action_list
       description: Returns a list of all the actions with runner parameters included.
       parameters:
@@ -514,7 +514,7 @@ paths:
             $ref: '#/definitions/Error'
   /api/v1/actions/views/overview/{ref_or_id}:
     get:
-      operationId: st2api.controllers.v1.actionviews:overview_controller.get_one
+      operationId: st2api.controllers.v1.action_views:overview_controller.get_one
       description: |
         Get one action with runner parameters included.
       parameters:
@@ -543,7 +543,7 @@ paths:
             $ref: '#/definitions/Error'
   /api/v1/actions/views/entry_point/{ref_or_id}:
     get:
-      operationId: st2api.controllers.v1.actionviews:entry_point_controller.get_one
+      operationId: st2api.controllers.v1.action_views:entry_point_controller.get_one
       description: |
         Get code of the action's entry_point.
       parameters:
@@ -1305,7 +1305,7 @@ paths:
             $ref: '#/definitions/Error'
   /api/v1/executions/views/filters:
     get:
-      operationId: st2api.controllers.v1.executionviews:filters_controller.get_all
+      operationId: st2api.controllers.v1.execution_views:filters_controller.get_all
       x-permissions: execution_views_filters_list
       description: |
         Get a list of distinct values for the execution filters.
@@ -2690,7 +2690,7 @@ paths:
             $ref: '#/definitions/Error'
   /api/v1/rules/views:
     get:
-      operationId: st2api.controllers.v1.ruleviews:rule_view_controller.get_all
+      operationId: st2api.controllers.v1.rule_views:rule_view_controller.get_all
       x-permissions: rule_list
       description: Returns a list of all rules.
       parameters:
@@ -2760,7 +2760,7 @@ paths:
             $ref: '#/definitions/Error'
   /api/v1/rules/views/{ref_or_id}:
     get:
-      operationId: st2api.controllers.v1.ruleviews:rule_view_controller.get_one
+      operationId: st2api.controllers.v1.rule_views:rule_view_controller.get_one
       description: |
         Get one rule.
       parameters:
